requirements.txt
openai
streamlit
Pillow
requests
openai.error
AuthenticationError